Move Routes component into component directory
