+++ src/index.js
@@ -5,7 +5,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import rootReducer from './reducers/index';
-import Routes from './containers/Routes';
+import Routes from './components/Routes';
 
 const store = createStore(
   rootReducer,
